Rename all other instances with mapview prefix
