+++ microapps/MapViewGeometryEditorApp/app/src/androidTest/java/com/arcgismaps/toolkit/geometryeditorapp/ExampleInstrumentedTest.kt
@@ -16,7 +16,7 @@
  *
  */
 
-package com.arcgismaps.toolkit.geometryeditorapp
+package com.arcgismaps.toolkit.mapviewgeometryeditorapp
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
@@ -35,6 +35,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.arcgismaps.toolkit.geometryeditorapp", appContext.packageName)
+        assertEquals("com.arcgismaps.toolkit.mapviewgeometryeditorapp", appContext.packageName)
     }
 }
