+++ microapps/MapViewGeometryEditorApp/app/src/main/AndroidManifest.xml
@@ -30,12 +30,12 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.GeometryEditorApp"
+        android:theme="@style/Theme.MapViewGeometryEditorApp"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:theme="@style/Theme.GeometryEditorApp">
+            android:theme="@style/Theme.MapViewGeometryEditorApp">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
