+++ microapps/MapViewLocationDisplayApp/app/src/androidTest/java/com/arcgismaps/toolkit/locationdisplayapp/ExampleInstrumentedTest.kt
@@ -16,7 +16,7 @@
  *
  */
 
-package com.arcgismaps.toolkit.locationdisplayapp
+package com.arcgismaps.toolkit.mapviewlocationdisplayapp
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
@@ -35,6 +35,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.arcgismaps.toolkit.locationdisplayapp", appContext.packageName)
+        assertEquals("com.arcgismaps.toolkit.mapviewlocationdisplayapp", appContext.packageName)
     }
 }
