+++ microapps/MapViewLocationDisplayApp/app/src/main/AndroidManifest.xml
@@ -32,12 +32,12 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.LocationDisplayApp"
+        android:theme="@style/Theme.MapViewLocationDisplayApp"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:theme="@style/Theme.LocationDisplayApp">
+            android:theme="@style/Theme.MapViewLocationDisplayApp">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
