+++ microapps/MapViewLocationDisplayApp/app/src/main/java/com/arcgismaps/toolkit/locationdisplayapp/ui/theme/Color.kt
@@ -16,7 +16,7 @@
  *
  */
 
-package com.arcgismaps.toolkit.locationdisplayapp.ui.theme
+package com.arcgismaps.toolkit.mapviewlocationdisplayapp.ui.theme
 
 import androidx.compose.ui.graphics.Color
