+++ microapps/MapViewLocationDisplayApp/app/src/main/res/values/strings.xml
@@ -17,5 +17,5 @@
   -->
 
 <resources>
-    <string name="app_name">LocationDisplayApp</string>
+    <string name="app_name">MapViewLocationDisplayApp</string>
 </resources>
