+++ microapps/MapViewLocationDisplayApp/app/src/test/java/com/arcgismaps/toolkit/locationdisplayapp/ExampleUnitTest.kt
@@ -16,7 +16,7 @@
  *
  */
 
-package com.arcgismaps.toolkit.locationdisplayapp
+package com.arcgismaps.toolkit.mapviewlocationdisplayapp
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
